quick access - introduce and reuse editor navigation base type
