--- src/vs/editor/contrib/quickAccess/quickAccess.ts
@@ -1,104 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { IQuickAccessProvider } from 'vs/platform/quickinput/common/quickAccess';
-import { IEditor } from 'vs/editor/common/editorCommon';
-import { IModelDeltaDecoration, OverviewRulerLane } from 'vs/editor/common/model';
-import { IRange } from 'vs/editor/common/core/range';
-import { themeColorFromId } from 'vs/platform/theme/common/themeService';
-import { overviewRulerRangeHighlight } from 'vs/editor/common/view/editorColorRegistry';
-import { IQuickPick, IQuickPickItem } from 'vs/platform/quickinput/common/quickInput';
-import { CancellationToken } from 'vs/base/common/cancellation';
-import { IDisposable } from 'vs/base/common/lifecycle';
-import { Event } from 'vs/base/common/event';
-
-interface IEditorLineDecoration {
-	rangeHighlightId: string;
-	overviewRulerDecorationId: string;
-}
-
-/**
- * A reusable quick access provider for the editor with support for adding decorations.
- */
-export abstract class AbstractEditorQuickAccessProvider implements IQuickAccessProvider {
-
-	/**
-	 * Subclasses to provide an event when the active editor control changes.
-	 */
-	abstract readonly onDidActiveTextEditorControlChange: Event<void>;
-
-	/**
-	 * Subclasses to provide the current active editor control.
-	 */
-	abstract activeTextEditorControl: IEditor | undefined;
-
-	/**
-	 * Subclasses to implement the quick access picker.
-	 */
-	abstract provide(picker: IQuickPick<IQuickPickItem>, token: CancellationToken): IDisposable;
-
-
-	//#region Decorations Utils
-
-	private rangeHighlightDecorationId: IEditorLineDecoration | undefined = undefined;
-
-	protected addDecorations(editor: IEditor, range: IRange): void {
-		editor.changeDecorations(changeAccessor => {
-
-			// Reset old decorations if any
-			const deleteDecorations: string[] = [];
-			if (this.rangeHighlightDecorationId) {
-				deleteDecorations.push(this.rangeHighlightDecorationId.overviewRulerDecorationId);
-				deleteDecorations.push(this.rangeHighlightDecorationId.rangeHighlightId);
-
-				this.rangeHighlightDecorationId = undefined;
-			}
-
-			// Add new decorations for the range
-			const newDecorations: IModelDeltaDecoration[] = [
-
-				// highlight the entire line on the range
-				{
-					range,
-					options: {
-						className: 'rangeHighlight',
-						isWholeLine: true
-					}
-				},
-
-				// also add overview ruler highlight
-				{
-					range,
-					options: {
-						overviewRuler: {
-							color: themeColorFromId(overviewRulerRangeHighlight),
-							position: OverviewRulerLane.Full
-						}
-					}
-				}
-			];
-
-			const [rangeHighlightId, overviewRulerDecorationId] = changeAccessor.deltaDecorations(deleteDecorations, newDecorations);
-
-			this.rangeHighlightDecorationId = { rangeHighlightId, overviewRulerDecorationId };
-		});
-	}
-
-	protected clearDecorations(editor: IEditor): void {
-		const rangeHighlightDecorationId = this.rangeHighlightDecorationId;
-		if (rangeHighlightDecorationId) {
-			editor.changeDecorations(changeAccessor => {
-				changeAccessor.deltaDecorations([
-					rangeHighlightDecorationId.overviewRulerDecorationId,
-					rangeHighlightDecorationId.rangeHighlightId
-				], []);
-			});
-
-			this.rangeHighlightDecorationId = undefined;
-		}
-	}
-
-	//#endregion
-}
