+++ src/vs/editor/standalone/browser/quickAccess/standaloneGotoSymbolQuickAccess.ts
@@ -13,13 +13,13 @@ import { Event } from 'vs/base/common/event';
 
 export class StandaloneGotoSymbolQuickAccessProvider extends AbstractGotoSymbolQuickAccessProvider {
 
-	readonly onDidActiveTextEditorControlChange = Event.None;
+	protected readonly onDidActiveTextEditorControlChange = Event.None;
 
 	constructor(@ICodeEditorService private readonly editorService: ICodeEditorService) {
 		super();
 	}
 
-	get activeTextEditorControl() {
+	protected get activeTextEditorControl() {
 		return withNullAsUndefined(this.editorService.getFocusedCodeEditor());
 	}
 }
